+++ src/component/playlist/playlist.js
@@ -1,0 +1,20 @@
+import './playlist.css'
+import Track from '../track/track'
+import arrOfTracks from '../data'
+
+
+export default function PlayList() {
+  return (
+    <div className="content__playlist playlist">
+      {arrOfTracks.map((item) => (
+        <Track
+          title={item.title}
+          additionTitle={item.additionTitle}
+          author={item.author}
+          album={item.album}
+          time={item.time}
+        />
+      ))}
+    </div>
+  )
+}
